# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ICICLE AI Vector Service",
    "description": "Vector storage and retrieval service for the ICICLE AI tenant. Clients provide\ntheir own pre-computed embeddings; the service handles storage, filtered search\nand reranking.\n\n**Collections** are broad domains (e.g. `biology`, `chemistry`), each backed by\nits own Qdrant collection and HNSW index. **Topics** are optional sub-categories\nwithin a collection (e.g. `human`, `plant`).\n\n**User isolation.** Every request is authenticated with a Tapis access token via\nthe `X-Tapis-Token` header, and `user_id` is taken from the token's\n`tapis/username` claim — never from the request body. Collections are physically\nshared, but every read, write, search and delete is filtered by `user_id`, so\nusers never see or affect each other's data.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/retrieve": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Vector similarity search",
        "operationId": "retrieve_v1_retrieve_post",
        "parameters": [
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      }
    },
    "/v1/rerank/methods": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "List available rerank methods",
        "description": "What this deployment can actually do.\n\n``available`` is false for 'cross_encoder' when the optional model\ndependencies were not installed in the image.",
        "operationId": "rerank_methods_v1_rerank_methods_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RerankMethodList"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      }
    },
    "/v1/rerank": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Rerank search results",
        "operationId": "rerank_v1_rerank_post",
        "parameters": [
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RerankRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RerankResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetadataFilter": {
        "properties": {
          "conditions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Conditions",
            "description": "Key-value pairs to match against the metadata payload field"
          }
        },
        "type": "object",
        "title": "MetadataFilter",
        "description": "Filter results by metadata fields.\n\nEach key-value pair matches against the nested ``metadata`` payload field.\nAll conditions are ANDed together. Values can be:\n  - a string/number/bool for exact match\n  - a list for \"any of\" match\n\nKeys are namespaced under ``metadata.`` when translated to Qdrant, so this\ncannot be used to address a top-level payload field such as ``user_id``."
      },
...
      "RerankMethodInfo": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "mmr",
              "cosine_rescore",
              "cross_encoder"
            ],
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "'vector' (embedding math) or 'cross-encoder' (model)."
          },
          "available": {
            "type": "boolean",
            "title": "Available",
            "description": "False when the method's dependencies are missing from this deployment."
          },
          "requires_query_text": {
            "type": "boolean",
            "title": "Requires Query Text"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "kind",
          "available",
          "requires_query_text",
          "description"
        ],
        "title": "RerankMethodInfo"
      },
      "RerankMethodList": {
        "properties": {
          "methods": {
            "items": {
              "$ref": "#/components/schemas/RerankMethodInfo"
            },
            "type": "array",
            "title": "Methods"
          },
          "default_model": {
            "type": "string",
            "title": "Default Model"
          },
          "allowed_models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Models"
          }
        },
        "type": "object",
        "required": [
          "methods",
          "default_model",
          "allowed_models"
        ],
        "title": "RerankMethodList"
      },
      "RerankRequest": {
        "properties": {
          "query_embedding": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Query Embedding"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "filter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetadataFilter"
              },
              {
                "type": "null"
              }
            ]
          },
          "query_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Text",
            "description": "Raw query string. Required for method='cross_encoder', which scores the query against each candidate's text. Ignored by the vector-only methods ('mmr', 'cosine_rescore')."
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 10
          },
          "fetch_k": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Fetch K",
            "default": 50
          },
          "method": {
            "type": "string",
            "enum": [
              "mmr",
              "cosine_rescore",
              "cross_encoder"
            ],
            "title": "Method",
            "default": "mmr"
          },
          "rerank_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Model",
            "description": "Cross-encoder model to use. Must be one of the models the deployment allows (see GET /v1/rerank/methods). Defaults to the service's configured model. Only used by method='cross_encoder'."
          },
          "lambda": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Lambda",
            "description": "MMR relevance/diversity trade-off. Only used by method='mmr'.",
            "default": 0.7
          }
        },
        "type": "object",
        "required": [
          "query_embedding",
          "collection"
        ],
        "title": "RerankRequest"
      },
      "RerankResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "method": {
            "type": "string",
            "enum": [
              "mmr",
              "cosine_rescore",
              "cross_encoder"
            ],
            "title": "Method"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Reranker model used. Null for the vector-only methods."
          },
          "top_k": {
            "type": "integer",
            "title": "Top K"
          },
          "fetch_k": {
            "type": "integer",
            "title": "Fetch K"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RerankedItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "method",
          "top_k",
          "fetch_k",
          "results"
        ],
        "title": "RerankResponse"
      },
      "RerankedItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "chunks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Chunks"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "rerank_score": {
            "type": "number",
            "title": "Rerank Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "score",
          "collection",
          "chunks",
          "metadata",
          "rerank_score"
        ],
        "title": "RerankedItem",
        "description": "A result carrying both the vector score and the reranker's score.\n\n``score`` stays the original cosine similarity from the vector search so\nclients can compare the two rankings; ``rerank_score`` is what the ordering\nis actually based on."
      },
      "ResultItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "chunks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Chunks"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "score",
          "collection",
          "chunks",
          "metadata"
        ],
        "title": "ResultItem"
      },
      "RetrieveRequest": {
        "properties": {
          "query_embedding": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Query Embedding"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "filter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetadataFilter"
              },
              {
                "type": "null"
              }
            ]
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "query_embedding",
          "collection"
        ],
        "title": "RetrieveRequest"
      },
      "RetrieveResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "top_k",
          "results"
        ],
        "title": "RetrieveResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "TapisToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tapis-Token",
        "description": "ICICLE AI tenant Tapis access token. The username is read from the tapis/username claim and used to isolate your data."
      }
    }
  },
  "servers": [
    {
      "url": "https://icicleaivecserver.pods.icicleai.tapis.io",
      "description": "Production (ICICLE AI Tapis pod)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
...
}